# jouko-api-spec/jaxrs-spec-generated/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Jouko API spec",
    "version" : "0.0.4",
    "title" : "Jouko API"
  },
  "basePath" : "/v1",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json;charset=utf-8" ],
  "produces" : [ "application/json;charset=utf-8" ],
  "security" : [ {
    "bearer" : [ ]
  } ],
  "paths" : {
...
    "/admin/users/{userId}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Retreive user",
        "operationId" : "retrieveUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The id of the user",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Users" ],
        "summary" : "Update user",
        "operationId" : "updateUser",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "The id of the user",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "The user to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/keycloakUsers/{keycloakId}" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "Get an user by keycloak id",
        "operationId" : "getUserByKeycloakId",
        "parameters" : [ {
          "name" : "keycloakId",
          "in" : "path",
          "description" : "The keycloak id of the user to look up",
          "required" : true,
          "type" : "string",
          "format" : "uuid"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/User"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
    "/keycloakUsers" : {
      "get" : {
        "tags" : [ "Users" ],
        "summary" : "List users",
        "operationId" : "listKeycloakUsers",
        "parameters" : [ {
          "name" : "token",
          "in" : "query",
          "description" : "Keycloak token",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/User"
              }
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Unauthorized"
            }
          },
          "500" : {
            "description" : "Internal server error",
            "schema" : {
              "$ref" : "#/definitions/InternalServerError"
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions" : {
    "bearer" : {
      "type" : "apiKey",
      "name" : "Authorization",
      "in" : "header"
    }
  },
  "definitions" : {
    "Unauthorized" : {
      "type" : "object",
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "The error message"
        }
      }
    },
    "InternalServerError" : {
      "type" : "object",
      "required" : [ "message" ],
      "properties" : {
        "message" : {
          "type" : "string",
          "description" : "The error message"
        }
      }
    },
    "User" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64",
          "description" : "User id"
        },
        "keycloakId" : {
          "type" : "string",
          "format" : "uuid",
          "description" : "The keycloak id of the user"
        },
        "firstName" : {
          "type" : "string",
          "description" : "The firstname id of the user"
        },
        "lastName" : {
          "type" : "string",
          "description" : "The lastname id of the user"
        },
        "email" : {
          "type" : "string",
          "description" : "The email id of the user"
        },
        "username" : {
          "type" : "string",
          "description" : "The username id of the user"
        }
      }
    },
...
  }
}